security(deploy): verify the deployed daemon by content, not by size (#1172)
The reuse gate in materialize_daemon_exe_to was
`dest.len() == source.len()`. Size is not integrity: a file of the right
length with the wrong bytes was reused and then executed as the daemon,
on every spawn, at a well-known path, with no content check anywhere
between materialization and Command::new.

A test asserted exactly that behaviour — it wrote a same-length sentinel
over the destination and required it to be reused untouched — so the
vulnerability was pinned in place. It is now inverted, and its doc
comment records what it used to assert and why that was wrong.

The fix needs no build-time digest, contrary to the issue's sketch: the
daemon IS a copy of the running CLI (multi-call binary, argv[0]
dispatch), so the source is its own reference. Compare blake3 of source
and dest, streamed in chunks rather than reading 20 MB into memory, with
the size check kept in front as a cheap reject. A read error answers
"not equal" and routes into copy-and-rename — the safe direction, since
the alternative is executing a file we could not verify.

The deploy directory is now created owner-only through
ensure_dir_private: anything that can write there chooses what the CLI
executes. Tightened or refused, it emits insecure_deploy_dir, following
the insecure_socket_dir contract from #1171.

runtime.rs crossed the 1.5K-LOC hard guard with this, so it splits into
runtime/mod.rs (should we have a daemon, and which one) plus
runtime/deploy.rs (where does the binary live, and is it the one we
think), with re-exports leaving every caller path unchanged.
